* Fix mupen64 across devices. * Fix multiple extension data errors * Fix alsa library package oops.

+++ packages/emulators/standalone/mupen64plus-sa/mupen64plus-sa-rsp-hle/package.mk
@@ -1,5 +1,7 @@
 # SPDX-License-Identifier: GPL-2.0-or-later
 # Copyright (C) 2019-present Shanti Gilbert (https://github.com/shantigilbert)
+# Copyright (C) 2023 Nicholas Ricciuti (rishooty@gmail.com)
+# Copyright (C) 2023 Fewtarius (fewtarius@jelos.org)
 
 PKG_NAME="mupen64plus-sa-rsp-hle"
 PKG_VERSION="f22dc143771f1a0784c7d62977722a68fa0bdf85"
@@ -13,42 +15,49 @@ PKG_TOOLCHAIN="manual"
 
 case ${DEVICE} in
   AMD64|RK3588|S922X)
-    PKG_DEPENDS_TARGET="mupen64plus-sa-simplecore"
+    PKG_DEPENDS_TARGET+=" mupen64plus-sa-simplecore"
   ;;
 esac
 
-if [ ! "${OPENGL}" = "no" ]; then
+if [ "${OPENGL_SUPPORT}" = "yes" ]
+then
   PKG_DEPENDS_TARGET+=" ${OPENGL} glu libglvnd"
-fi
-
-if [ "${OPENGLES_SUPPORT}" = yes ]; then
+else
   PKG_DEPENDS_TARGET+=" ${OPENGLES}"
 fi
 
 make_target() {
-  case ${ARCH} in
-    arm|aarch64)
-      export HOST_CPU=aarch64
-      BINUTILS="$(get_build_dir binutils)/.aarch64-libreelec-linux-gnueabi"
-      export USE_GLES=1
-    ;;
-    x86_64)
-      export HOST_CPU=x86_64
-      export USE_GLES=0
-    ;;
-  esac
-  export APIDIR=${SYSROOT_PREFIX}/usr/local/include/mupen64plus
-  export SDL_CFLAGS="-I${SYSROOT_PREFIX}/usr/include/SDL2 -pthread"
+  if [ "${OPENGL_SUPPORT}" = "yes" ]
+  then
+    export USE_GLES=0
+  else
+    export USE_GLES=1
+  fi
+
+  export HOST_CPU=${TARGET_ARCH} \
+         VFP_HARD=1 \
+         NEW_DYNAREC=1 \
+         V=1 \
+         VC=0 \
+         OSD=0
+
+  export BINUTILS="$(get_build_dir binutils)/.${TARGET_NAME}"
+  export APIDIR=$(get_build_dir mupen64plus-sa-core)/src/api
+  export SDL_CFLAGS="-I${SYSROOT_PREFIX}/usr/include/SDL2 -pthread -D_REENTRANT"
   export SDL_LDLIBS="-lSDL2_net -lSDL2"
   export CROSS_COMPILE="${TARGET_PREFIX}"
-  export V=1
-  export VC=0
+
   make -C projects/unix clean
   make -C projects/unix all ${PKG_MAKE_OPTS_TARGET}
   cp ${PKG_BUILD}/projects/unix/mupen64plus-rsp-hle.so ${PKG_BUILD}/projects/unix/mupen64plus-rsp-hle-base.so
-  export APIDIR=${SYSROOT_PREFIX}/usr/local/include/simple64
-  make -C projects/unix all ${PKG_MAKE_OPTS_TARGET}
-  cp ${PKG_BUILD}/projects/unix/mupen64plus-rsp-hle.so ${PKG_BUILD}/projects/unix/mupen64plus-rsp-hle-simple.so
+
+  case ${DEVICE} in
+    AMD64|RK3588|S922X)
+      export APIDIR=$(get_build_dir mupen64plus-sa-simple64)/src/api
+      make -C projects/unix all ${PKG_MAKE_OPTS_TARGET}
+      cp ${PKG_BUILD}/projects/unix/mupen64plus-rsp-hle.so ${PKG_BUILD}/projects/unix/mupen64plus-rsp-hle-simple.so
+    ;;
+  esac
 }
 
 makeinstall_target() {
@@ -57,10 +66,12 @@ makeinstall_target() {
   UPLUGINDIR=${ULIBDIR}/mupen64plus
   mkdir -p ${UPLUGINDIR}
   cp ${PKG_BUILD}/projects/unix/mupen64plus-rsp-hle-base.so ${UPLUGINDIR}/mupen64plus-rsp-hle.so
-  #${STRIP} ${UPLUGINDIR}/mupen64plus-rsp-hle.so
   chmod 0644 ${UPLUGINDIR}/mupen64plus-rsp-hle.so
-  cp ${PKG_BUILD}/projects/unix/mupen64plus-rsp-hle-simple.so ${UPLUGINDIR}
-  #${STRIP} ${UPLUGINDIR}/mupen64plus-rsp-hle-simple.so
-  chmod 0644 ${UPLUGINDIR}/mupen64plus-rsp-hle-simple.so
+
+  if [ -e "${PKG_BUILD}/projects/unix/mupen64plus-rsp-hle-simple.so" ]
+  then
+    cp ${PKG_BUILD}/projects/unix/mupen64plus-rsp-hle-simple.so ${UPLUGINDIR}
+    chmod 0644 ${UPLUGINDIR}/mupen64plus-rsp-hle-simple.so
+  fi
 }
